+++ src/CloneMerge.cpp
@@ -8,6 +8,7 @@ struct CloneMerge : CloneModuleBase {
 
   enum ParamId {
     CLONE_PARAM,
+    GROUP_PARAM,
     PARAMS_LEN
   };
   enum InputId {
@@ -32,6 +33,7 @@ struct CloneMerge : CloneModuleBase {
       configInput(MONO_INPUTS+i, string::f("Mono %d", i + 1));
       configLight(MONO_LIGHTS+i*2, string::f("Input %d cloned indicator", i + 1))->description = "yellow = OK, red = Error";
     }
+    configSwitch<FixedSwitchQuantity>(GROUP_PARAM, 0.f, 1.f, 0.f, "Output grouping", {"Input channel", "Input set"});
     configOutput(POLY_OUTPUT, "Poly");
     lightDivider.setDivision(44);
   }
@@ -52,13 +54,22 @@ struct CloneMerge : CloneModuleBase {
     }
     int goodIns = ins > maxIns ? maxIns : ins;
 
-    int channel=0;
-    for (int i=0; i<goodIns; i+=1) {
-      float val = inputs[MONO_INPUTS + i].getVoltage();
-      for (int c=0; c<clones; c++)
-        outputs[POLY_OUTPUT].setVoltage(val, channel++);
+    if (params[GROUP_PARAM].getValue()){ //group by input set
+      for (int i=0; i<goodIns; i++) {
+        float val = inputs[MONO_INPUTS+i].getVoltage();
+        for (int c=0, o=i; c<clones; c++, o+=goodIns)
+          outputs[POLY_OUTPUT].setVoltage(val, o);
+      }
     }
-    outputs[POLY_OUTPUT].setChannels(channel);
+    else{ // group by input channel
+      for (int i=0, o=0; i<goodIns; i+=1) {
+        float val = inputs[MONO_INPUTS + i].getVoltage();
+        for (int c=0; c<clones; c++)
+          outputs[POLY_OUTPUT].setVoltage(val, o++);
+      }
+    }
+    outputs[POLY_OUTPUT].setChannels(goodIns * clones);
+    
     processExpander(clones, goodIns);
     
     if (lightDivider.process()) {
@@ -88,6 +99,9 @@ struct CloneMergeWidget : CloneModuleWidget {
       addInput(createInputCentered<MonoPort>(Vec(x,y), module, CloneMerge::MONO_INPUTS + i));
       addChild(createLightCentered<SmallLight<YellowRedLight<>>>(Vec(x+dy*0.5f, y-dy*0.3f), module, CloneMerge::MONO_LIGHTS + i*2));
     }
+
+    addParam(createLockableParamCentered<GroupSwitch>(Vec(6.f,311.5f), module, CloneMerge::GROUP_PARAM));
+
     y+=dy*0.33f;
     addOutput(createOutputCentered<PolyPort>(Vec(x,y), module, CloneMerge::POLY_OUTPUT));
   }