+++ src/CloneModule.hpp
@@ -5,6 +5,7 @@ struct CloneModule : VenomModule {
 
   #define EXPANDER_PORTS 4
   enum ExpParamId {
+    ENUMS(EXP_GROUP_PARAM,4),
     EXP_PARAMS_LEN
   };
   enum ExpInputId {
@@ -26,7 +27,6 @@ struct CloneModule : VenomModule {
 struct CloneModuleBase : CloneModule {
   
   void processExpander(int clones, int goodCh){
-    int expanderChannels[EXPANDER_PORTS]{};
     Module* expander = getRightExpander().module;
     expander = expander
              && !expander->isBypassed() 
@@ -34,20 +34,24 @@ struct CloneModuleBase : CloneModule {
               ? expander 
               : NULL;
     if (!expander) return;
-    for (int i=0; i<EXPANDER_PORTS; i++){
-      if (expander->outputs[EXP_POLY_OUTPUT+i].isConnected())
-        expanderChannels[i] = std::max(expander->inputs[EXP_POLY_INPUT+i].getChannels(),1);
-    }
-    for (int c=0; c<goodCh; c++) {
-      for (int e=0; e<EXPANDER_PORTS; e++){
-        float val = expander->inputs[EXP_POLY_INPUT+e].getPolyVoltage(c);
-        for (int j=0, ec=c*clones; j<clones; j++, ec++)
-          expander->outputs[EXP_POLY_OUTPUT+e].setVoltage(val, ec);
-      }
-    }
     int outCnt = clones * goodCh;
-    for (int e=0; e<EXPANDER_PORTS; e++){
-      expander->outputs[EXP_POLY_OUTPUT+e].setChannels( expanderChannels[e] ? outCnt : 0 );
+    for (int p=0; p<EXPANDER_PORTS; p++){
+      if (expander->params[EXP_GROUP_PARAM+p].getValue()) { // group by input set
+        for (int c=0, o=0; c<clones; c++){
+          for (int i=0; i<goodCh; i++, o++){
+            float val = expander->inputs[EXP_POLY_INPUT+p].getPolyVoltage(i);
+            expander->outputs[EXP_POLY_OUTPUT+p].setVoltage(val, o);
+          }
+        }
+      }
+      else { // group by individual input channel
+        for (int i=0, o=0; i<goodCh; i++){
+          float val = expander->inputs[EXP_POLY_INPUT+p].getPolyVoltage(i);
+          for (int c=0; c<clones; c++, o++)
+            expander->outputs[EXP_POLY_OUTPUT+p].setVoltage(val, o);
+        }
+      }
+      expander->outputs[EXP_POLY_OUTPUT+p].setChannels(outCnt);
     }
   }
 
@@ -89,6 +93,14 @@ struct CloneModuleBase : CloneModule {
 };
 
 struct CloneModuleWidget : VenomWidget {
+
+  struct GroupSwitch : GlowingSvgSwitchLockable {
+    GroupSwitch() {
+      addFrame(Svg::load(asset::plugin(pluginInstance,"res/smallYellowButtonSwitch.svg")));
+      addFrame(Svg::load(asset::plugin(pluginInstance,"res/smallBlueButtonSwitch.svg")));
+    }
+  };
+
   void appendContextMenu(Menu* menu) override {
     menu->addChild(new MenuSeparator());
     Module* expander = module->rightExpander.module;